requirements.txt
pandas
requests
djangorestframework
python-decouple
django-cors-headers